Moving typed array tests into TypedArray.rs
