+++ crates/js-sys/tests/wasm/Array.rs
@@ -1,7 +1,5 @@
 use js_sys::*;
 use std::iter::FromIterator;
-use std::mem;
-use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
 use wasm_bindgen::JsValue;
 use wasm_bindgen_test::*;
@@ -536,192 +534,4 @@ fn array_inheritance() {
     assert!(array.is_instance_of::<Array>());
     assert!(array.is_instance_of::<Object>());
     let _: &Object = array.as_ref();
-}
-
-#[wasm_bindgen(module = "tests/wasm/Array.js")]
-extern "C" {
-    fn populate_array(arr: JsValue, start: JsValue, len: JsValue) -> JsValue;
-}
-
-fn test_array_view_mut_raw<ElemT: std::cmp::PartialEq + std::fmt::Debug, ArrT>(
-    sut: unsafe fn(*mut ElemT, usize) -> ArrT,
-    u8ToElem: fn(u8) -> ElemT,
-    arrToJsValue: fn(ArrT) -> JsValue,
-) {
-    let start: u8 = 10;
-    let len: usize = 32;
-    let end: u8 = start + len as u8;
-    let mut buffer: Vec<ElemT> = Vec::new();
-    buffer.reserve(len);
-    unsafe {
-        let array: ArrT = sut(buffer.as_mut_ptr(), len);
-        populate_array(
-            arrToJsValue(array),
-            JsValue::from(start),
-            JsValue::from(len as u32),
-        );
-        buffer.set_len(len);
-    }
-    let expected: Vec<ElemT> = (start..end).map(u8ToElem).collect();
-    assert_eq!(buffer, expected)
-}
-
-fn u8Toi8_unsafe(x: u8) -> i8 {
-    x as i8
-}
-
-#[wasm_bindgen_test]
-fn Int8Array_view_mut_raw() {
-    test_array_view_mut_raw(
-        js_sys::Int8Array::view_mut_raw,
-        u8Toi8_unsafe,
-        JsValue::from,
-    );
-}
-
-#[wasm_bindgen_test]
-fn Int16Array_view_mut_raw() {
-    test_array_view_mut_raw(js_sys::Int16Array::view_mut_raw, i16::from, JsValue::from);
-}
-
-#[wasm_bindgen_test]
-fn Int32Array_view_mut_raw() {
-    test_array_view_mut_raw(js_sys::Int32Array::view_mut_raw, i32::from, JsValue::from);
-}
-
-#[wasm_bindgen_test]
-fn Uint8Array_view_mut_raw() {
-    test_array_view_mut_raw(js_sys::Uint8Array::view_mut_raw, u8::from, JsValue::from);
-}
-
-#[wasm_bindgen_test]
-fn Uint8ClampedArray_view_mut_raw() {
-    test_array_view_mut_raw(
-        js_sys::Uint8ClampedArray::view_mut_raw,
-        u8::from,
-        JsValue::from,
-    );
-}
-
-#[wasm_bindgen_test]
-fn Uint16Array_view_mut_raw() {
-    test_array_view_mut_raw(js_sys::Uint16Array::view_mut_raw, u16::from, JsValue::from);
-}
-
-#[wasm_bindgen_test]
-fn Uint32Array_view_mut_raw() {
-    test_array_view_mut_raw(js_sys::Uint32Array::view_mut_raw, u32::from, JsValue::from);
-}
-
-#[wasm_bindgen_test]
-fn Float32Array_view_mut_raw() {
-    test_array_view_mut_raw(js_sys::Float32Array::view_mut_raw, f32::from, JsValue::from);
-}
-
-#[wasm_bindgen_test]
-fn Float64Array_view_mut_raw() {
-    test_array_view_mut_raw(js_sys::Float64Array::view_mut_raw, f64::from, JsValue::from);
-}
-
-fn test_js_array_copy_to_maybe_uninit<ElemT: std::cmp::PartialEq + std::fmt::Debug, JsArrT>(
-    sut: unsafe fn(&JsArrT, &mut mem::MaybeUninit<Vec<ElemT>>) -> (),
-    sliceToJsArray: fn(&[ElemT]) -> JsArrT,
-    u8ToElem: fn(u8) -> ElemT,
-) {
-    let start: u8 = 10;
-    let len: usize = 32;
-    let end: u8 = start + len as u8;
-    let expected: Vec<ElemT> = (start..end).map(u8ToElem).collect();
-
-    let js_arr: JsArrT = sliceToJsArray(expected.as_slice());
-
-    let mut mvec: mem::MaybeUninit<Vec<ElemT>> = mem::MaybeUninit::uninit();
-
-    unsafe {
-        sut(&js_arr, &mut mvec);
-
-        assert_eq!(*mvec.as_mut_ptr(), expected);
-        assert_eq!(mvec.assume_init(), expected);
-    }
-}
-
-#[wasm_bindgen_test]
-fn Int8Array_copy_to_maybe_uninit() {
-    test_js_array_copy_to_maybe_uninit(
-        js_sys::Int8Array::copy_to_maybe_uninit,
-        |xs: &[i8]| js_sys::Int8Array::from(xs),
-        u8Toi8_unsafe,
-    );
-}
-
-#[wasm_bindgen_test]
-fn Int16Array_copy_to_maybe_uninit() {
-    test_js_array_copy_to_maybe_uninit(
-        js_sys::Int16Array::copy_to_maybe_uninit,
-        |xs: &[i16]| js_sys::Int16Array::from(xs),
-        i16::from,
-    );
-}
-
-#[wasm_bindgen_test]
-fn Int32Array_copy_to_maybe_uninit() {
-    test_js_array_copy_to_maybe_uninit(
-        js_sys::Int32Array::copy_to_maybe_uninit,
-        |xs: &[i32]| js_sys::Int32Array::from(xs),
-        i32::from,
-    );
-}
-
-#[wasm_bindgen_test]
-fn Uint8Array_copy_to_maybe_uninit() {
-    test_js_array_copy_to_maybe_uninit(
-        js_sys::Uint8Array::copy_to_maybe_uninit,
-        |xs: &[u8]| js_sys::Uint8Array::from(xs),
-        u8::from,
-    );
-}
-
-#[wasm_bindgen_test]
-fn Uint8ClampedArray_copy_to_maybe_uninit() {
-    test_js_array_copy_to_maybe_uninit(
-        js_sys::Uint8ClampedArray::copy_to_maybe_uninit,
-        |xs: &[u8]| js_sys::Uint8ClampedArray::from(xs),
-        u8::from,
-    );
-}
-
-#[wasm_bindgen_test]
-fn Uint16Array_copy_to_maybe_uninit() {
-    test_js_array_copy_to_maybe_uninit(
-        js_sys::Uint16Array::copy_to_maybe_uninit,
-        |xs: &[u16]| js_sys::Uint16Array::from(xs),
-        u16::from,
-    );
-}
-
-#[wasm_bindgen_test]
-fn Uint32Array_copy_to_maybe_uninit() {
-    test_js_array_copy_to_maybe_uninit(
-        js_sys::Uint32Array::copy_to_maybe_uninit,
-        |xs: &[u32]| js_sys::Uint32Array::from(xs),
-        u32::from,
-    );
-}
-
-#[wasm_bindgen_test]
-fn Float32Array_copy_to_maybe_uninit() {
-    test_js_array_copy_to_maybe_uninit(
-        js_sys::Float32Array::copy_to_maybe_uninit,
-        |xs: &[f32]| js_sys::Float32Array::from(xs),
-        f32::from,
-    );
-}
-
-#[wasm_bindgen_test]
-fn Float64Array_copy_to_maybe_uninit() {
-    test_js_array_copy_to_maybe_uninit(
-        js_sys::Float64Array::copy_to_maybe_uninit,
-        |xs: &[f64]| js_sys::Float64Array::from(xs),
-        f64::from,
-    );
 }
