+++ crates/js-sys/tests/wasm/TypedArray.js
@@ -1,4 +1,4 @@
-// Used for `Array.rs` tests
+// Used for `TypedArray.rs` tests
 exports.populate_array =  function(arr, start, len) {
   for (i = 0; i < len; i++) {
     arr[i] = start + i;
